refactor: split record.go
