+++ internal/pkg/metadb/record/record.go
@@ -15,30 +15,12 @@
 package record
 
 import (
-	"fmt"
-	"reflect"
-
 	"cloud.google.com/go/datastore"
 	"github.com/google/uuid"
 	pb "github.com/googleforgames/open-saves/api"
 	"github.com/googleforgames/open-saves/internal/pkg/metadb/timestamps"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
-
-// PropertyValue is an internal representation of the user-defined property.
-// See the Open Saves API definition for details.
-type PropertyValue struct {
-	Type         pb.Property_Type
-	IntegerValue int64
-	StringValue  string
-	BooleanValue bool
-}
-
-// PropertyMap represents user-defined custom properties.
-type PropertyMap map[string]*PropertyValue
-
-// Assert PropertyMap implements PropertyLoadSave.
-var _ datastore.PropertyLoadSaver = new(PropertyMap)
 
 // Record represents a Open Saves record in the metadata database.
 // See the Open Saves API definition for details.
@@ -60,42 +42,6 @@ type Record struct {
 // Assert Record implements both PropertyLoadSave and KeyLoader.
 var _ datastore.PropertyLoadSaver = new(Record)
 var _ datastore.KeyLoader = new(Record)
-
-// ToProto converts the struct to a proto.
-func (p *PropertyValue) ToProto() *pb.Property {
-	ret := &pb.Property{
-		Type: p.Type,
-	}
-	switch p.Type {
-	case pb.Property_BOOLEAN:
-		ret.Value = &pb.Property_BooleanValue{BooleanValue: p.BooleanValue}
-	case pb.Property_INTEGER:
-		ret.Value = &pb.Property_IntegerValue{IntegerValue: p.IntegerValue}
-	case pb.Property_STRING:
-		ret.Value = &pb.Property_StringValue{StringValue: p.StringValue}
-	}
-	return ret
-}
-
-// NewPropertyValueFromProto creates a new Property instance from a proto.
-// Passing nil returns a zero-initialized Property.
-func NewPropertyValueFromProto(proto *pb.Property) *PropertyValue {
-	if proto == nil {
-		return new(PropertyValue)
-	}
-	ret := &PropertyValue{
-		Type: proto.Type,
-	}
-	switch proto.Type {
-	case pb.Property_BOOLEAN:
-		ret.BooleanValue = proto.GetBooleanValue()
-	case pb.Property_INTEGER:
-		ret.IntegerValue = proto.GetIntegerValue()
-	case pb.Property_STRING:
-		ret.StringValue = proto.GetStringValue()
-	}
-	return ret
-}
 
 const externalBlobPropertyName = "ExternalBlob"
 
@@ -168,101 +114,4 @@ func NewRecordFromProto(p *pb.Record) *Record {
 			UpdatedAt: p.GetUpdatedAt().AsTime(),
 		},
 	}
-}
-
-// NewPropertyMapFromProto creates a new Property instance from a proto.
-// Passing nil returns an empty map.
-func NewPropertyMapFromProto(proto map[string]*pb.Property) PropertyMap {
-	if proto == nil {
-		return make(PropertyMap)
-	}
-	ret := make(PropertyMap)
-	for k, v := range proto {
-		ret[k] = NewPropertyValueFromProto(v)
-	}
-	return ret
-}
-
-// ToProto converts the struct to a proto.
-func (m *PropertyMap) ToProto() map[string]*pb.Property {
-	// This may seem wrong, but m is a pointer to a map, which is also a
-	// nullable reference type.
-	if m == nil || *m == nil {
-		return nil
-	}
-	ret := make(map[string]*pb.Property)
-	for k, v := range *m {
-		ret[k] = v.ToProto()
-	}
-	return ret
-}
-
-// Save implements the Datastore PropertyLoadSaver interface and converts
-// PropertyMap to a slice of datastore Properties.
-func (m *PropertyMap) Save() ([]datastore.Property, error) {
-	var ps []datastore.Property
-	if m == nil {
-		return ps, nil
-	}
-	for name, value := range *m {
-		switch value.Type {
-		case pb.Property_BOOLEAN:
-			ps = append(ps, datastore.Property{
-				Name:  name,
-				Value: value.BooleanValue,
-			})
-		case pb.Property_INTEGER:
-			ps = append(ps, datastore.Property{
-				Name:  name,
-				Value: value.IntegerValue,
-			})
-		case pb.Property_STRING:
-			ps = append(ps, datastore.Property{
-				Name:  name,
-				Value: value.StringValue,
-			})
-		default:
-			return nil,
-				fmt.Errorf(
-					"error storing property, unknown type: name=[%s], type=[%s]",
-					name, value.Type.String())
-		}
-	}
-	return ps, nil
-}
-
-// Load implements the Datastore PropertyLoadSaver interface and converts
-// individual properties to PropertyMap.
-func (m *PropertyMap) Load(ps []datastore.Property) error {
-	if ps == nil || len(ps) == 0 {
-		// No custom properties
-		return nil
-	}
-	if m == nil || *m == nil {
-		// I don't think this should happen because the Datastore Go client
-		// always zero-initializes the target before calling Load.
-		return fmt.Errorf("PropertyMap.Load was called on nil")
-	}
-
-	for _, v := range ps {
-		var newValue = new(PropertyValue)
-		t := reflect.TypeOf(v.Value)
-		switch t.Kind() {
-		case reflect.Bool:
-			newValue.Type = pb.Property_BOOLEAN
-			newValue.BooleanValue = v.Value.(bool)
-		case reflect.Int64:
-			newValue.Type = pb.Property_INTEGER
-			newValue.IntegerValue = v.Value.(int64)
-		case reflect.String:
-			newValue.Type = pb.Property_STRING
-			newValue.StringValue = v.Value.(string)
-		default:
-			return fmt.Errorf(
-				"error loading property, unknown type: name=[%s], type=[%s]",
-				v.Name, t.Name())
-		}
-		(*m)[v.Name] = newValue
-	}
-	return nil
 }
