+++ internal/pkg/metadb/record/record_test.go
@@ -26,109 +26,6 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 )
 
-func TestPropertyValue_ToProto(t *testing.T) {
-	booleanProperty := &PropertyValue{Type: pb.Property_BOOLEAN, BooleanValue: true}
-	booleanExpected := &pb.Property{
-		Type:  pb.Property_BOOLEAN,
-		Value: &pb.Property_BooleanValue{BooleanValue: true},
-	}
-	assert.Equal(t, booleanExpected, booleanProperty.ToProto())
-
-	integerProperty := &PropertyValue{Type: pb.Property_INTEGER, IntegerValue: 42}
-	integerExpected := &pb.Property{
-		Type:  pb.Property_INTEGER,
-		Value: &pb.Property_IntegerValue{IntegerValue: 42},
-	}
-	assert.Equal(t, integerExpected, integerProperty.ToProto())
-
-	stringProperty := &PropertyValue{Type: pb.Property_STRING, StringValue: "string value"}
-	stringExpected := &pb.Property{
-		Type:  pb.Property_STRING,
-		Value: &pb.Property_StringValue{StringValue: "string value"},
-	}
-	assert.Equal(t, stringExpected, stringProperty.ToProto())
-}
-
-func TestPropertyValue_NewPropertyFromProtoNil(t *testing.T) {
-	expected := new(PropertyValue)
-	actual := NewPropertyValueFromProto(nil)
-	assert.NotNil(t, expected, actual)
-	assert.Equal(t, expected, actual)
-}
-
-func TestPropertyValue_NewPropertyFromProto(t *testing.T) {
-	booleanProto := &pb.Property{
-		Type:  pb.Property_BOOLEAN,
-		Value: &pb.Property_BooleanValue{BooleanValue: true},
-	}
-	booleanExpected := &PropertyValue{Type: pb.Property_BOOLEAN, BooleanValue: true}
-	assert.Equal(t, booleanExpected, NewPropertyValueFromProto(booleanProto))
-
-	integerProto := &pb.Property{
-		Type:  pb.Property_INTEGER,
-		Value: &pb.Property_IntegerValue{IntegerValue: 42},
-	}
-	integerExpected := &PropertyValue{Type: pb.Property_INTEGER, IntegerValue: 42}
-	assert.Equal(t, integerExpected, NewPropertyValueFromProto(integerProto))
-
-	stringProto := &pb.Property{
-		Type:  pb.Property_STRING,
-		Value: &pb.Property_StringValue{StringValue: "string value"},
-	}
-	stringExpected := &PropertyValue{Type: pb.Property_STRING, StringValue: "string value"}
-	assert.Equal(t, stringExpected, NewPropertyValueFromProto(stringProto))
-
-}
-
-func TestPropertyMap_NewPropertyMapFromProto(t *testing.T) {
-	// Handles nil
-	assert.Equal(t, make(PropertyMap), NewPropertyMapFromProto(nil))
-
-	proto := map[string]*pb.Property{
-		"boolean": {Type: pb.Property_BOOLEAN,
-			Value: &pb.Property_BooleanValue{BooleanValue: true}},
-		"int": {Type: pb.Property_INTEGER,
-			Value: &pb.Property_IntegerValue{IntegerValue: -50}},
-		"string": {Type: pb.Property_STRING,
-			Value: &pb.Property_StringValue{StringValue: "abc"},
-		},
-	}
-	expected := PropertyMap{
-		"boolean": {Type: pb.Property_BOOLEAN, BooleanValue: true},
-		"int":     {Type: pb.Property_INTEGER, IntegerValue: -50},
-		"string":  {Type: pb.Property_STRING, StringValue: "abc"},
-	}
-	actual := NewPropertyMapFromProto(proto)
-	if assert.NotNil(t, actual) {
-		assert.Equal(t, expected, actual)
-	}
-}
-
-func TestPropertyMap_ToProto(t *testing.T) {
-	// Handles nil
-	var nilMap PropertyMap
-	assert.Nil(t, nilMap.ToProto())
-
-	propertyMap := PropertyMap{
-		"boolean": {Type: pb.Property_BOOLEAN, BooleanValue: true},
-		"int":     {Type: pb.Property_INTEGER, IntegerValue: -50},
-		"string":  {Type: pb.Property_STRING, StringValue: "abc"},
-	}
-	expected := map[string]*pb.Property{
-		"boolean": {Type: pb.Property_BOOLEAN,
-			Value: &pb.Property_BooleanValue{BooleanValue: true}},
-		"int": {Type: pb.Property_INTEGER,
-			Value: &pb.Property_IntegerValue{IntegerValue: -50}},
-		"string": {Type: pb.Property_STRING,
-			Value: &pb.Property_StringValue{StringValue: "abc"},
-		},
-	}
-	actual := propertyMap.ToProto()
-	if assert.NotNil(t, actual) {
-		assert.Equal(t, expected, actual)
-	}
-}
-
 func TestRecord_Save(t *testing.T) {
 	testBlob := []byte{0x24, 0x42, 0x11}
 	createdAt := time.Date(1992, 1, 15, 3, 15, 55, 0, time.UTC)
